Switch type of multiple 512* accounts to "disponibilites"
By default those accounts tend to hold monetary assets (that result
in a negative balance) and must be accounted that way.

The former type "emprunts_dettes_etablissements_credit" is only suited
for loans, where the account will typically have a positive balance.
Those should be handled in sub-accounts that will be configured with
this type.

issue3633
review3171003