Cleanup usage of reconcile field
Drop the field on all accounts except on customers/suppliers accounts.
Also drop the field on all view accounts.

Reconciliation is not required for any account but it's usually done
for customers and suppliers account. If the user want to do it for more
accounts, he can customize his accounts but by default we should not force
the user to reconcile more lines than what is usually done.

issue4468
review5931002