Switch account 58 into a real account
It's only a temporary account to host moves between other accounts
but it needs to be a real (non-view) account to be able to host moves. Its
balance should be zero most of the time so the classification doesn't
matter much but we'll classify it as other bank accounts since it usually
holds monetary assets which are in transit between two accounts.

issue4523
review8911002