Remove tax rule "Ventes intracommunautaires" and associated taxes
There is no need for these taxes: in case of sale of goods in europe excluding
VAT exemption, we have to account and declare this sale and the VAT as if the
sale was made in France.

TODO: Deactivate all those taxes, either in the XML or in account.py.